+++ Changelog.md
@@ -111,4 +111,6 @@
 |01.02.2023|M365CallFlowVisualizerV2.ps1|2.9.4|Rename -NoCache to -CacheResults as it's easier to not to think about a double negative value.|
 |03.02.2023|M365CallFlowVisualizerV2.ps1|2.9.5|Add support for Call Queue Welcome Text-To-Speech Greetings. Requires MicrosoftTeams PowerShell 4.9.3.|
 |12.02.2023|M365CallFlowVisualizerV2.ps1|2.9.6|Performance improvements (reduce number of Get-Cs*), Add parameter to show number type for CQ Agents, outbound calling Ids, Loop until a top level voice app is selected.|
-|13.02.2023|M365CallFlowVisualizerV2.ps1|2.9.6|Fix bug where transfer message was displayed in AA default/after hours call flow without IVR. Add user calling settings to nested holiday call flows.|
+|13.02.2023|M365CallFlowVisualizerV2.ps1|2.9.7|Fix bug where transfer message was displayed in AA default/after hours call flow without IVR. Add user calling settings to nested holiday call flows. Extend -MaxResults from 1000 to 9999.|
+|13.02.2023|Get-AutoAttendantDirectorySearchConfig.ps1|1.0.0|Initial development to read search scope configurations from AAs.|
+|13.02.2023|M365CallFlowVisualizerV2.ps1|2.9.8|Include directory search scope for AAs with voice menus.|